% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggrmodel.R
\name{aggrmodel}
\alias{aggrmodel}
\title{Fit Aggregated Model}
\usage{
aggrmodel(
  formula = NULL,
  data,
  market,
  Y = NULL,
  timeVar,
  groupVar,
  repVar,
  n_basis,
  cicleRep = FALSE,
  n_order = 4,
  covType = "Homog_Uniform",
  corType = "periodic",
  diffTol = 1e-06
)
}
\arguments{
\item{formula}{building...}

\item{data}{Dataset containing group, replicates (if any), time and aggregated signal}

\item{market}{Market data frame. MUST be a 3 column dataframe with the following order: Group, Type and Number of subjects}

\item{Y}{Dependent variable: aggregated signal}

\item{timeVar}{Name of time variable}

\item{groupVar}{Name of grouping variable}

\item{repVar}{Name of replicates variable}

\item{n_basis}{Number of basis functions for basis expansion}

\item{cicleRep}{Indicator TRUE/FALSE if replicates are cyclical}

\item{n_order}{Order of basis Splines (Default: 4)}

\item{covType}{Covariance functional type. One of "Homog_Uniform", "Homog" or "Heterog"}

\item{corType}{Correlation structure type. One of "periodic" or "exponential"}

\item{diffTol}{Tolerance of model covergence}
}
\value{
beta and sigma pars
}
\description{
Fit Aggregated Model
}
