% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodFunctions.R
\name{buildX}
\alias{buildX}
\title{Build design matrix to fit aggregated model}
\usage{
buildX(
  marketLong,
  nType,
  timeVec,
  n_basis,
  basis = "B-Splines",
  n_order = 4,
  timeVec2 = NULL,
  n_basis2 = NULL,
  basis2 = NULL,
  n_order2 = NULL
)
}
\arguments{
\item{marketLong}{Market indexed by group and rep}

\item{nType}{Number of subjects type}

\item{timeVec}{Vector of sampled times}

\item{n_basis}{Number of basis functions for basis expansion}

\item{basis}{Character indicating which basis: 'B-Splines' or 'Fourier'}

\item{n_order}{Order of basis splines (Default: 4)}

\item{timeVec2}{Vector of second functional component}

\item{n_basis2}{Number of basis function for timeVec2 expansion}

\item{basis2}{Character indicating which basis: 'B-Splines' or 'Fourier'}

\item{n_order2}{Order of B-Splines (Default:4)}
}
\value{
List of basis matrices for each group in market[,1]
}
\description{
Build design matrix to fit aggregated model
}
\examples{
## Create market example and time vector
myMkt = data.frame(group = rep(1:4, each=3),
                   type = rep(1:3, times=4),
                   mkt = sample(1:20, size=12, replace=TRUE))
myTimeVec = seq(0,1, length.out=12)

designList = buildX(myMkt, myTimeVec, n_basis = 5)
str(designList)
}
