% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarianceFunctions.R
\name{covMatrix}
\alias{covMatrix}
\title{Covariance matrix for aggregated model}
\usage{
covMatrix(
  market,
  group.name,
  type.name,
  mkt.name,
  timeVec,
  sigPar,
  tauPar,
  corPar,
  funcMtx = NULL,
  covType,
  corType = "periodic",
  nKnots = NULL,
  truncateDec = NULL
)
}
\arguments{
\item{market}{Market information (improve)}

\item{group.name}{empty}

\item{type.name}{empty}

\item{mkt.name}{empty}

\item{timeVec}{empty}

\item{sigPar}{improve}

\item{tauPar}{empty}

\item{corPar}{empty}

\item{funcMtx}{empty}

\item{covType}{Any of 'Homog_uniform', 'Homog' or 'Heterog'.}

\item{corType}{Any of 'periodic' (default) or 'exponential'.}

\item{nKnots}{empty}

\item{truncateDec}{empty}
}
\value{
A list of C covariance matrices of size \code{ncol(funcMtx)}
}
\description{
Covariance matrix for aggregated model
}
\details{
building...
}
\examples{
set.seed(2019)

## Create market
mkt = data.frame(group = rep(1:3, each=2),
                 type = rep(1:2, times = 3),
                 value = sample(1:20, 6))

myTimevec = seq(0,1, length.out = 4)
mySigPar = matrix(c(2,3), ncol=2)
myTauPar = matrix(c(.2, .2), ncol=2)
myCorPar = matrix(c(8, 12), ncol=2)
myFuncMtx = matrix(runif(8), nrow = 4, ncol=2)

## Homogeneous example
homogMtx = covMatrix(market = mkt, group.name = 'group', type.name = 'type', mkt.name = 'value', timeVec = myTimevec, sigPar = mySigPar, tauPar = myTauPar, corPar = myCorPar, covType = 'Homog', corType = 'periodic')

}
