% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarianceFunctions.R
\name{createCorBase}
\alias{createCorBase}
\title{Create auxiliary correlation base matrix}
\usage{
createCorBase(timeVec)
}
\arguments{
\item{timeVec}{Vector of time}
}
\value{
Square correlation matrix of size \code{length(unique(timeVec))}
}
\description{
Create auxiliary correlation base matrix
}
\examples{
createCorBase(seq(0,1, length.out=12))
}
