% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuFunctions.R
\name{createSimuData}
\alias{createSimuData}
\title{Simulate aggregated data}
\usage{
createSimuData(
  nGroups = 10,
  market = NULL,
  market_range = 1:100,
  sigPar = c(2, 2, 2),
  tauPar = c(0.5, 0.5, 0.5),
  corPar = c(13, 13, 13),
  seed = NULL,
  tempPar = 1
)
}
\arguments{
\item{nGroups}{Number of groups to be simulated}

\item{market}{Optional. Insert the market as a data frame with the columns 'group', 'type' and 'num'}

\item{market_range}{If market not inserted, the range to be sampled}

\item{sigPar}{Parameters of sigma: vector of length 3}

\item{tauPar}{Parameters of tau: vector of length 3}

\item{corPar}{Parameters of correlation: vector of length 3}

\item{seed}{Optional. Seed to be used for simulation}

\item{tempPar}{Default:1. If you do not want temperature effect, set temPar=0}
}
\value{
Return a data frame with 20 replicates and 3 types of consumer observed in 48 times.
}
\description{
Simulate aggregated data
}
\examples{
dd = createSimuData()
dd = createSimuData(nGroups=4)
}
