% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarianceFunctions.R
\name{createVarMtx}
\alias{createVarMtx}
\title{Functional variance matrix}
\usage{
createVarMtx(functionalVec, sigPar, tauPar)
}
\arguments{
\item{functionalVec}{Vector containing values for the matrix diagonal}

\item{sigPar}{Scale parameter (see details)}

\item{tauPar}{Non-negative power parameter (see details)}
}
\value{
A square functional variace matrix with size = \code{length(functionalVec)}
}
\description{
Functional variance matrix
}
\details{
The functional variance matrix is a matrix with diagonal of elements:

\deqn{v(t) = \sigma \, (\eta(t))^{-\tau}}
}
