% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarianceFunctions.R
\name{expCorMtx}
\alias{expCorMtx}
\title{Exponential correlation matrix}
\usage{
expCorMtx(timeVec, corPar, truncateDec = NULL)
}
\arguments{
\item{timeVec}{Vector of time}

\item{corPar}{Covariance parameter}

\item{truncateDec}{Decimal to be truncated. Default is none.}
}
\value{
Square correlation matrix of size \code{length(unique(timeVec))} of type exponential.
}
\description{
Exponential correlation matrix
}
\details{
Create a covariance matrix of type \emph{exponential} with elements

\deqn{r(s,t | \theta) = exp \left\{-2\theta\, \left(\frac{|t-s|}{T}\right) \right\}.}
}
\examples{
myTime <- seq(0,1, length.out = 12)
expCorMtx(myTime, corPar = 8, truncateDec = 4)
}
