% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initsFunctions.R
\name{get_inits}
\alias{get_inits}
\title{get_inits}
\usage{
get_inits(
  X,
  I,
  data,
  C,
  covType,
  market,
  sigPar_init,
  corPar_init,
  betaCov_init,
  truncateDec,
  n_basis,
  n_basis_cov,
  t,
  n_order,
  basisFunc
)
}
\arguments{
\item{X}{Design matrix}

\item{I}{Number of replicates}

\item{C}{Number of subject types}

\item{covType}{Covariance structure type}

\item{market}{Model market}

\item{sigPar_init}{inital values for sigma parameter}

\item{corPar_init}{Initial value for correlation parameters}

\item{betaCov_init}{inital values for variance functional}

\item{truncateDec}{Decimal to be truncated}

\item{n_basis}{Number of basis function in model fit}

\item{n_basis_cov}{Number of basis function in heterogeneous case}

\item{t}{time vector}

\item{n_order}{Order of basis}

\item{basisFunc}{Basis function type}

\item{y}{Dependent variable}

\item{tauPar_init}{Initial value for covariance parameter}
}
\value{
An object with initial values
}
\description{
Aux function to get initial values in aggrmodel function
}
\author{
Gabriel Franco
}
