% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggrmodel_cluster.R
\name{getClusterInitials}
\alias{getClusterInitials}
\title{Fit multiple simple models to obtain the clustering with minor SRE}
\usage{
getClusterInitials(data, market, n_knots, n_order, n_cluster, n_trials, bFunc)
}
\arguments{
\item{data}{Must be an ordered data frame by rep > group > time with theses names}

\item{market}{A data frame with columns group, type and num}

\item{n_knots}{Number of knots to be fitted on simple model}

\item{n_order}{Order of basis Splines (Default: 4)}

\item{n_cluster}{Number of grouping clusters}

\item{n_trials}{Number of random clustering configurations to be fitted}

\item{bFunc}{Basis function to be used: "B-Splines" or "Fourier"}
}
\value{
A list containing: data set with "group" and "cluster" indicating the fit with minor squared root error and an lm object of the chosen clustering configuration
}
\description{
Fit multiple simple models to obtain the clustering with minor SRE
}
