% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodFunctions.R
\name{gradLK}
\alias{gradLK}
\title{Create gradient function approximation via PRACMA package}
\usage{
gradLK(
  x,
  dataWrap,
  mktWrap,
  sCovWrap,
  covWrap,
  corWrap,
  betaWrap,
  designWrap,
  nCons,
  nBasisCov,
  nOrderCov,
  verbWrap,
  positive = FALSE,
  truncateDec = NULL
)
}
\arguments{
\item{x}{parameters to be evaluated}

\item{dataWrap}{dataset}

\item{mktWrap}{market dataset}

\item{sCovWrap}{Sample covariance matrix}

\item{covWrap}{covariance structure}

\item{corWrap}{correlation structure}

\item{betaWrap}{beta parameter}

\item{designWrap}{Design matrix}

\item{nCons}{number of types of consumers}

\item{nBasisCov}{number of basis functions for functional variance}

\item{nOrderCov}{order of basis functions for functional variance}

\item{verbWrap}{TRUE/FALSE prints likelihood values during optimization}

\item{positive}{Use positive restriction?}

\item{truncateDec}{Integer: Decimal to be truncated in exponential functional}
}
\value{

}
\description{
Create gradient function approximation via PRACMA package
}
