% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{market}
\alias{market}
\title{Market for simulated data}
\format{A data frame with 24 obs. of 4 variables:
\describe{
\item{Group}{Grouping variable with 8 different groups}
\item{Cluster}{Clustering variable with 2 clusters}
\item{Type}{Subject type}
\item{Num}{Number of subjects of this type in this}
}}
\source{
Simulated data
}
\usage{
market
}
\description{
A dataset containing market information for simulated aggregated data.
}
\keyword{datasets}
