% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarianceFunctions.R
\name{periodicCorMtx}
\alias{periodicCorMtx}
\title{Periodic correlation matrix}
\usage{
periodicCorMtx(timeVec, corPar, truncateDec = NULL)
}
\arguments{
\item{timeVec}{Vector of time}

\item{corPar}{Covariance parameter}

\item{truncateDec}{Decimal to be truncated. Default is none.}
}
\value{
Square correlation matrix of size \code{length(unique(timeVec))} of type periodic.
}
\description{
Periodic correlation matrix
}
\details{
Create a covariance matrix of type \emph{periodic} with elements

\deqn{r(s,t | \theta) = exp \left\{-2\theta\, sen^2\left(\pi\frac{(t-s)}{T}\right) \right\}.}
}
\examples{
myTime <- seq(0,1, length.out = 12)
periodicCorMtx(myTime, corPar = 8, truncateDec = 4)
}
