% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelCheckFunctions.R
\name{plot.aggrmodel_cluster}
\alias{plot.aggrmodel_cluster}
\title{Plot mean curves of Aggregated Model with clusters}
\usage{
\method{plot}{aggrmodel_cluster}(object, scales = "fixed")
}
\arguments{
\item{object}{aggrmodel_cluster object}

\item{scales}{Character value indicating if scales should be 'fixed' or 'free' in the ggplot}
}
\value{
A ggplot2 object
}
\description{
Plot mean curves of Aggregated Model with clusters
}
\examples{
library(dplyr)
data = simuData \%>\%
  select(group=Group, rep=Rep, time=Time, y=Load)
df=data
mkt = market
mkt$Cluster=NULL
colnames(mkt) = c('group','type','num')

## Go for a walk after run the code below
fitCluster = aggrmodel_cluster(data = df, market=mkt, Y = 'y', timeVar =
'time',groupVar = 'group', repVar = 'rep', n_basis = 7,n_basis_cov = NULL,
n_cluster = 2,n_trials = 1000, n_order = 4, corType = 'periodic', verbose=TRUE)

plot(fitCluster)
}
