% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelCheckFunctions.R
\name{plotFitted}
\alias{plotFitted}
\title{Plot fitted values over observed values}
\usage{
plotFitted(object, obsAlpha = 0.5, predColor = "magenta", scales = "fixed")
}
\arguments{
\item{object}{aggrmodel object fit}

\item{obsAlpha}{Numeric between 0 and 1 indicating observed curves transparency}

\item{predColor}{Character or numeric indicating fitted curves color}
}
\description{
Plot fitted values over observed values
}
