% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelCheckFunctions.R
\name{plotMC}
\alias{plotMC}
\title{Plot fit's mean curve}
\usage{
plotMC(object)
}
\arguments{
\item{object}{aggrmodel object fit}
}
\value{
ggplot plot object
}
\description{
Plot fit's mean curve
}
