% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodFunctions.R
\name{Q_function}
\alias{Q_function}
\title{Function to be maximized in aggrmodel_cluster function}
\usage{
Q_function(data, sigmaList, xbetaList, probTab, B)
}
\arguments{
\item{data}{Data Frame with 4 columns in the following order: Group, Replicates, Time, Signal}

\item{sigmaList}{List of covariance matrices for each group}

\item{xbetaList}{List of estimated signal for each group}

\item{probTab}{Data frame of expected probabilities from E-Step containing a group variable, replicate variable and cluster probabilities}

\item{B}{Number of grouping clusters}
}
\value{
A scalar likelihood value
}
\description{
Function to be maximized in aggrmodel_cluster function
}
