% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodFunctions.R
\name{Q_wrapper}
\alias{Q_wrapper}
\title{Wrapper for M-Step function evaluation}
\usage{
Q_wrapper(
  covPar,
  data,
  market,
  betaPar,
  piPar,
  pTab,
  B,
  t,
  K,
  I,
  J,
  basisFunction,
  n_order
)
}
\arguments{
\item{covPar}{Covariance parameters}

\item{data}{Data Frame with 4 columns in the following order: Group, Replicates, Time, Signal}

\item{market}{Data Frame with 4 columns in the following order: Group, Replicates, Time, Signal}

\item{betaPar}{Mean curves parameters}

\item{piPar}{Probabilities parameters}

\item{pTab}{Data frame of expected probabilities from E-Step containing a group variable, replicate variable and cluster probabilities}

\item{B}{Number of grouping clusters}

\item{t}{Time vector}

\item{K}{Number of basis functions}

\item{I}{Number of replicates}

\item{J}{Number of groups}

\item{basisFunction}{Character indicating which basis: 'B-Splines' (default) or 'Fourier'}

\item{n_order}{Order of basis Splines (Default: 4)}
}
\value{

}
\description{
Wrapper for M-Step function evaluation
}
