% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggrmodel_cluster.R
\name{simpleAggrmodel}
\alias{simpleAggrmodel}
\title{Fit a simple aggregated model}
\usage{
simpleAggrmodel(data, market, n_type, n_basis, n_order, basisFunction)
}
\arguments{
\item{data}{Dataset containing group, replicates (if any), time and aggregated signal}

\item{market}{Market data frame. MUST be a 3 column dataframe with the following order: Group, Type and Number of subjects}

\item{n_basis}{Number of basis functions for basis expansion}

\item{n_order}{Order of basis Splines (Default: 4)}

\item{basisFunction}{Character indicating which basis: 'B-Splines' (default) or 'Fourier'}
}
\value{
lm object
}
\description{
Fit a simple aggregated model
}
