% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simuData}
\alias{simuData}
\title{Simulated clustered aggregated signal}
\format{A data frame with 9600 obs. of 7 variables:
\describe{
  \item{group}{Grouping variable with 8 different groups}
  \item{rep}{Replication variable with 4 different replicates}
  \item{temperature}{Temperature information}
  \item{cluster}{Clustering variable with 2 clusters}
  \item{time}{Time variable with 48 samples times}
  \item{y}{Simulated dependent aggregated variable}
  \item{y_real}{Non noisy dependent aggregated variable}
}

The parameters and market used for simulation are in attributes.}
\source{
Simulated data by createSimuData function
}
\usage{
simuData
}
\description{
A dataset containing 10 groups divided into 2 clusters (B1=4,B2=6) with 20 replications each.
}
\keyword{datasets}
