% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simuTemperature}
\alias{simuTemperature}
\title{Simulated Temperature for simulation analysis}
\format{A data frame with 960 obs. of 3 variables:
\describe{
\item{Time}{Observed times in 30 minutes interval}
\item{Day}{Day flag: from 1 to 20}
\item{Temp}{Temperature: day 1 to 10 is a Brazilian summer scenario and day 11 to 20 a Brazilian winter scenario}
}}
\usage{
simuTemperature
}
\description{
A dataset containing temperature for 20 days observed in 30 minutes intervals
}
\keyword{datasets}
