% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggrmodel.R
\name{aggrmodel}
\alias{aggrmodel}
\title{Fit Aggregated Model}
\usage{
aggrmodel(
  formula = NULL,
  data,
  market,
  Y = NULL,
  timeVar,
  groupVar,
  repVar,
  n_basis,
  n_basis_cov = NULL,
  basisFunction = "B-Splines",
  cicleRep = FALSE,
  n_order = 4,
  covType = "Homog_Uniform",
  corType = "periodic",
  corPar_init = 20,
  tauPar_init = 0.5,
  optimMethod = "L-BFGS-B",
  truncateDec = NULL,
  verbose = FALSE,
  diffTol = 1e-06
)
}
\arguments{
\item{formula}{building...}

\item{data}{Dataset containing group, replicates (if any), time and aggregated signal}

\item{market}{Market data frame. MUST be a 3 column dataframe with the following order: Group, Type and Number of subjects}

\item{Y}{Dependent variable: aggregated signal}

\item{timeVar}{Name of time variable}

\item{groupVar}{Name of grouping variable}

\item{repVar}{Name of replicates variable}

\item{n_basis}{Number of basis functions for basis expansion}

\item{n_basis_cov}{Number of basis functions for variance functional expansion}

\item{basisFunction}{Character indicating which basis: 'B-Splines' or 'Fourier'}

\item{cicleRep}{Indicator TRUE/FALSE if replicates are cyclical}

\item{n_order}{Order of basis Splines (Default: 4)}

\item{covType}{Covariance functional type. One of "Homog_Uniform", "Homog" or "Heterog"}

\item{corType}{Correlation structure type. One of "periodic" or "exponential"}

\item{corPar_init}{Numeric: Initial value for correlation parameters (default:20)}

\item{tauPar_init}{Numeric: Initial value for expoent parameters of complete covariance (default:0.5)}

\item{verbose}{TRUE/FALSE prints likelihood values during optimization}

\item{diffTol}{Tolerance of model covergence}
}
\value{
An aggrmodel object
}
\description{
Fit Aggregated Model
}
\examples{
df = subset(simuData, Cluster==1)
mkt = subset(market, Cluster==1)
mkt = subset(mkt, select=-Cluster)

aggrFit = aggrmodel(data = df, market = mkt, Y = Load, timeVar = Time, groupVar = Group, repVar = Rep, n_basis = 7)
}
