% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodFunctions.R
\name{buildX}
\alias{buildX}
\title{Build design matrix to fit aggregated model}
\usage{
buildX(market, timeVec, n_basis, basis = "B-Splines", n_order = 4)
}
\arguments{
\item{market}{Market data frame. MUST be a 3 column dataframe with the following order: Group, Type and Number of subjects}

\item{timeVec}{Vector of sampled times}

\item{n_basis}{Number of basis functions for basis expansion}

\item{basis}{Character indicating which basis: 'B-Splines' or 'Fourier'}

\item{n_order}{Order of basis splines (Default: 4)}
}
\value{
List of basis matrices for each group in market[,1]
}
\description{
Build design matrix to fit aggregated model
}
\examples{
## Create market example and time vector
myMkt = data.frame(group = rep(1:4, each=3),
                   type = rep(1:3, times=4),
                   mkt = sample(1:20, size=12, replace=TRUE))
myTimeVec = seq(0,1, length.out=12)

designList = buildX(myMkt, myTimeVec, n_basis = 5)
str(designList)
}
