% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuFunctions.R
\name{createSimuData}
\alias{createSimuData}
\title{Simulate aggregated data}
\usage{
createSimuData(
  B1 = 4,
  B2 = 6,
  nRep = 20,
  market = NULL,
  market_range = 5:25,
  sigPar = c(2, 2, 2, 4, 4, 4),
  tauPar = c(0.5, 0.5, 0.5, 0.4, 0.4, 0.4),
  corPar = c(4, 4, 4, 6, 6, 6),
  nu1 = NULL,
  nu2 = NULL,
  seed = NULL,
  tempPar = 1
)
}
\arguments{
\item{B1}{Number of groups from cluster 1}

\item{B2}{Number of groups from cluster 2}

\item{nRep}{Number of replicates}

\item{market}{Optional. Insert the market as a data frame with the columns 'group', 'type' and 'num'}

\item{market_range}{If market not inserted, the range to be sampled}

\item{sigPar}{Parameters of sigma: vector of length 3}

\item{tauPar}{Parameters of tau: vector of length 3}

\item{corPar}{Parameters of correlation: vector of length 3}

\item{nu1}{Functional of variance for cluster 1 (must be of length N)}

\item{nu2}{Functional of variance for cluster 2 (must be of length N)}

\item{seed}{Optional. Seed to be used for simulation}

\item{tempPar}{Default:1. If you do not want temperature effect, set temPar=0}
}
\value{
Return a data frame with nRep replicates and 3 types of consumer observed in 48 times. Parameters and market are returned as attributes.
}
\description{
Simulate aggregated data
}
\examples{
dd = createSimuData()
dd = createSimuData(B1=4, B2=8)
}
