% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodFunctions.R
\name{logLikelihood}
\alias{logLikelihood}
\title{Compute log-likelihood for aggregated model}
\usage{
logLikelihood(data, muVec, covMtxList)
}
\arguments{
\item{data}{Data Frame with 4 columns in the following order: Group, Replicates, Time, Signal}

\item{muVec}{Vector \eqn{X\beta} of fitted values}

\item{covMtxList}{List of covariance matrices for each group}
}
\value{
Log-likehood value
}
\description{
Compute log-likelihood for aggregated model
}
