% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodFunctions.R
\name{loglikWrapper}
\alias{loglikWrapper}
\title{Wrapper for log-likelihood evaluation}
\usage{
loglikWrapper(
  pars,
  dataWrap,
  mktWrap,
  covWrap,
  corWrap,
  betaWrap,
  designListWrap,
  nCons,
  nBasisCov,
  nOrderCov,
  verbWrap,
  truncateDec = NULL
)
}
\arguments{
\item{pars}{parameters to be evaluated in optim}

\item{dataWrap}{dataset}

\item{mktWrap}{market dataset}

\item{covWrap}{covariance structure}

\item{corWrap}{correlation structure}

\item{betaWrap}{beta parameter}

\item{designListWrap}{list of design matrices for each group}

\item{nCons}{number of types of consumers}

\item{nBasisCov}{number of basis functions for functional variance}

\item{nOrderCov}{order of basis functions for functional variance}

\item{verbWrap}{TRUE/FALSE prints likelihood values during optimization}

\item{truncateDec}{Integer: Decimal to be truncated in exponential functional}
}
\description{
Wrapper for log-likelihood evaluation
}
