% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelCheckFunctions.R
\name{plotMC}
\alias{plotMC}
\title{Plot fit's mean curve}
\usage{
plotMC(object, scales = "fixed")
}
\arguments{
\item{object}{aggrmodel object fit}

\item{scales}{Character value indicating if scales should be 'fixed' or 'free' in the ggplot}
}
\value{
ggplot plot object
}
\description{
Plot fit's mean curve
}
