% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggrmodel_cluster.R
\name{simpleAggrmodel}
\alias{simpleAggrmodel}
\title{Fit a simple aggregated model}
\usage{
simpleAggrmodel(data, market, n_basis, n_order, basisFunction)
}
\arguments{
\item{data}{}
}
\value{
lm object
}
\description{
Fit a simple aggregated model
}
