% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/simuFunctions.R
\docType{data}
\name{simuData}
\alias{simuData}
\title{Simulated clustered aggregated signal}
\format{A data frame with 1536 obs. of 8 variables:
\describe{
  \item{Group}{Grouping variable with 8 different groups}
  \item{C1}{Market information: customer of type C1}
  \item{C2}{Market information: customer of type C2}
  \item{C3}{Market information: customer of type C3}
  \item{Rep}{Replication variable with 4 different replicates}
  \item{Cluster}{Clustering variable with 2 clusters}
  \item{Time}{Time variable with 48 samples times}
  \item{Load}{Simulated dependent aggregated variable}
}}
\source{
Simulated data
}
\usage{
simuData

simuData(
  nGroups = 10,
  market = NULL,
  market_range = 1:100,
  sigPar = c(2, 2, 2),
  tauPar = c(0.5, 0.5, 0.5),
  corPar = c(13, 13, 13),
  seed = NULL,
  tempPar = 1
)
}
\arguments{
\item{nGroups}{Number of groups to be simulated}

\item{market}{Optional. Insert the market as a data frame with the columns 'group', 'type' and 'num'}

\item{market_range}{If market not inserted, the range to be sampled}

\item{sigPar}{Parameters of sigma: vector of length 3}

\item{tauPar}{Parameters of tau: vector of length 3}

\item{corPar}{Parameters of correlation: vector of length 3}

\item{seed}{Optional. Seed to be used for simulation}

\item{tempPar}{Default:1. If you do not want temperature effect, set temPar=0}
}
\value{
Return a data frame with 20 replicates and 3 types of consumer observed in 48 times.
}
\description{
A dataset containing 8 groups divided into 2 clusters with 4 replications each.
}
\examples{
dd = simuData()
dd = simuData(nGroups=4)
}
\keyword{datasets}
