% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simuData}
\alias{simuData}
\title{Simulated clustered aggregated signal}
\format{A data frame with 1536 obs. of 8 variables:
\describe{
  \item{Group}{Grouping variable with 8 different groups}
  \item{C1}{Market information: customer of type C1}
  \item{C2}{Market information: customer of type C2}
  \item{C3}{Market information: customer of type C3}
  \item{Rep}{Replication variable with 4 different replicates}
  \item{Cluster}{Clustering variable with 2 clusters}
  \item{Time}{Time variable with 48 samples times}
  \item{Load}{Simulated dependent aggregated variable}
}}
\source{
Simulated data
}
\usage{
simuData
}
\description{
A dataset containing 8 groups divided into 2 clusters with 4 replications each.
}
\keyword{datasets}
